:memo: :: 리프레쉬 싱글톤으로 변경 및 리프레쉬 실패시 예외처리
