+++ app/src/main/java/com/g3c1/aide/feature_splash/presentation/ui/SplashActivity.kt
@@ -2,6 +2,7 @@ package com.g3c1.aide.feature_splash.presentation.ui
 
 import android.content.Intent
 import android.os.Bundle
+import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.lifecycle.lifecycleScope
@@ -10,7 +11,7 @@ import com.g3c1.aide.feature_account.presentation.ui.AccountActivity
 import com.g3c1.aide.feature_account.presentation.utils.TokenType
 import com.g3c1.aide.feature_splash.presentation.ui.screen.SplashScreen
 import com.g3c1.aide.feature_store.presentation.ui.StoreActivity
-import com.g3c1.aide.remote.utils.TokenInterceptor
+import com.g3c1.aide.remote.utils.token_handler.TokenInterceptor
 import com.g3c1.aide.ui.theme.Orange
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import dagger.hilt.android.AndroidEntryPoint
@@ -31,12 +32,16 @@ class SplashActivity : ComponentActivity() {
                 AideBossApplication.getInstance().getTokenManager().getTokenData(TokenType.REFRESH)
             )
             delay(1000)
-            startActivity(
-                Intent(
-                    this@SplashActivity,
-                    if (isRefreshExpired) AccountActivity::class.java else StoreActivity::class.java
+            if (isRefreshExpired != null) {
+                startActivity(
+                    Intent(
+                        this@SplashActivity,
+                        if (isRefreshExpired) AccountActivity::class.java else StoreActivity::class.java
+                    )
                 )
-            )
+            } else {
+                Toast.makeText(this@SplashActivity, "알수 없는 오류가 발생했습니다.", Toast.LENGTH_SHORT).show()
+            }
         }
     }
 }
